Move common sync errors into main mod file
